Using #tag to change current progress state
#progress:number can be used in dialogues to change the progress state. The progress variable will update upon reading the line.

The number in the tag can be 0, 1, 2, or 3.
For the ProgressManager, a function called  transitionToNextState(int i) handles the transition. each i of 0~3 will lead to different states, although not all 4 states are valid.
For progress state transition based on finding clues,
if the clue leads to Boss, i should be 0
if the clue leads to Doc, i should be 1
if the clue leads to Chef, i should be 2
if the clue leads to Boss, i should be 3.
And the #progress:0/1/2/3 has been written in each of the evidence ink files.
The police call and other state transitions need to be hard coded.